renames module with more proper name
